Asterisks removed
Quick fix for non-CMU entries that pass through eng_to_ipa, as indicated in https://github.com/Mitchellpkt/Immersion/issues/1

Just did a find & replace for asterisks, which leaves a bunch of unnecessary entries where the output is the same as the input. Will update once the issue is fixed correctly.